sort bookmarks

+++ app/src/main/java/com/foobnix/model/AppBookmark.java
@@ -48,7 +48,7 @@ public class AppBookmark implements MyPath.RelativePath {
         return p;
     }
 
-    public float getTime() {
+    public long getTime() {
         return t;
     }
